requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==1.11.3
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-reversion==2.0.13
djangorestframework==3.6.3
future==0.16.0
httplib2==0.11.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
Pillow==5.1.0
pytz==2018.4
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
XlsxWriter==1.0.5
xlwt==1.3.0
